simple c ffi
