+++ Cargo.toml
@@ -1,9 +1,24 @@
-[workspace]
-members = [
-    "tree_sitter_grep",
-    "tree_sitter_grep_plugin_core",
-    "filter_after_line_number",
-]
+[package]
+name = "tree-sitter-grep"
+version = "0.1.0"
+edition = "2021"
 
-[workspace.dependencies]
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+clap = { version = "4.3.0", features = ["derive"] }
+grep = "0.2.12"
+ignore = "0.4.20"
+libloading = "0.8.0"
+once_cell = "1.17.1"
+rayon = "1.7.0"
+regex = "1.8.2"
 tree-sitter = "0.20.10"
+tree-sitter-rust = "0.20.3"
+tree-sitter-typescript = "0.20.2"
+
+[lib]
+name = "tree_sitter_grep"
+
+[[bin]]
+name = "tree-sitter-grep"
