--- filter_after_line_number/Cargo.toml
@@ -1,12 +1,0 @@
-[package]
-name = "filter_after_line_number"
-version = "0.1.0"
-edition = "2021"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-[lib]
-crate-type = ["cdylib"]
-
-[dependencies]
-tree_sitter_grep_plugin_core = { path = "../tree_sitter_grep_plugin_core" }
-tree-sitter.workspace = true
