--- tree_sitter_grep/Cargo.toml
@@ -1,25 +1,0 @@
-[package]
-name = "tree-sitter-grep"
-version = "0.1.0"
-edition = "2021"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-clap = { version = "4.3.0", features = ["derive"] }
-grep = "0.2.12"
-ignore = "0.4.20"
-libloading = "0.8.0"
-once_cell = "1.17.1"
-rayon = "1.7.0"
-regex = "1.8.2"
-tree-sitter.workspace = true
-tree-sitter-rust = "0.20.3"
-tree-sitter-typescript = "0.20.2"
-tree_sitter_grep_plugin_core = { path = "../tree_sitter_grep_plugin_core" }
-
-[lib]
-name = "tree_sitter_grep"
-
-[[bin]]
-name = "tree-sitter-grep"
