--- tree_sitter_grep_plugin_core/Cargo.toml
@@ -1,12 +1,0 @@
-[package]
-name = "tree_sitter_grep_plugin_core"
-version = "0.0.1"
-edition = "2021"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-tree-sitter.workspace = true
-
-[build-dependencies]
-rustc_version = "0.4.0"
