--- tree_sitter_grep_plugin_core/build.rs
@@ -1,4 +1,0 @@
-fn main() {
-    let version = rustc_version::version().unwrap();
-    println!("cargo:rustc-env=RUSTC_VERSION={}", version);
-}
