--- tree_sitter_grep_plugin_core/src/lib.rs
@@ -1,33 +1,0 @@
-use tree_sitter::Node;
-
-// per https://adventures.michaelfbryan.com/posts/plugins-in-rust/
-
-pub trait Filterer: Send + Sync {
-    fn call(&self, node: &Node) -> bool;
-}
-
-pub struct PluginDeclaration {
-    pub rustc_version: &'static str,
-    pub core_version: &'static str,
-    pub register: unsafe extern "C" fn(&mut dyn PluginRegistrar, Option<&str>),
-}
-
-pub trait PluginRegistrar {
-    fn register_filterer(&mut self, filterer: Box<dyn Filterer>);
-}
-
-pub static CORE_VERSION: &str = env!("CARGO_PKG_VERSION");
-pub static RUSTC_VERSION: &str = env!("RUSTC_VERSION");
-
-#[macro_export]
-macro_rules! export_plugin {
-    ($register:expr) => {
-        #[doc(hidden)]
-        #[no_mangle]
-        pub static plugin_declaration: $crate::PluginDeclaration = $crate::PluginDeclaration {
-            rustc_version: $crate::RUSTC_VERSION,
-            core_version: $crate::CORE_VERSION,
-            register: $register,
-        };
-    };
-}
